+++ src/pages/_app.page.tsx
@@ -9,7 +9,7 @@ import Head from 'next/head'
 import { trpc } from '@/utils/trpc'
 import { server } from '@/mocks/server'
 import ErrorBoundary from '@/components/ErrorBoundary/ErrorBoundary'
-import { BASE_PATH } from '@/constants/paths'
+import { BASE_PATH } from '@/constants/appConstants'
 import { initFaro } from '@/libs/faro/faro'
 import { ToggleProvider } from '@/contexts/toggleContext'
 import { TestScenarioSelector } from '@/components/TestscenarioSelector/TestScenarioSelector'
