+++ src/utils/utils.ts
@@ -1,6 +1,6 @@
 import { formPageInverted } from '@/domain/formPages'
 import { FormPage } from '@/types/merOppfolgingForm'
-import { FORM_PATH } from '@/constants/paths'
+import { FORM_PATH } from '@/constants/appConstants'
 
 export function getFormUrl(form: FormPage): string {
   return `${FORM_PATH}/${formPageInverted[form]}`
